README.md
This project is of a Simple Chatbot. Here greeting function for the chatbot, allow it to respond to at least five basic questions including a farewell message. This can remember the previous interactions, recall and reference the context of the conversation. First the chatbot ask the user some basic questions then the user can start the conversation with the chatbot. About the error handling to address scenarios where the chatbot does  not understand the user's input there is a friendly response in such cases.